resample: Renamed resample_nipy -> resample and move to spinalcord
Former-commit-id: fd387c4cedee4806ce9e0a081d6b3cafffd05f55 [formerly a9a5b2db29b555f321d9ea377ac6564ebd666777 [formerly 5e724840c6ecc939ccb304b6a0ab512a531f493e]]
Former-commit-id: 02b5c3458e0a6e282aaf6967675a7263b36375ca
Former-commit-id: 27a7c15de08acc4b1027904066177733d48e2f1f
