+++ spinalcordtoolbox/deepseg_gm/deepseg_gm.py
@@ -34,7 +34,7 @@ except Exception as e:
 else:
     sys.stderr = original_stderr
 
-from spinalcordtoolbox.resample import nipy_resample
+from spinalcordtoolbox import resample
 from . import model
 
 
@@ -311,10 +311,10 @@ def segment_file(input_filename, output_filename,
     pixdim = nii_original.header["pixdim"][3]
     target_resample = "0.25x0.25x{:.5f}".format(pixdim)
 
-    nii_resampled = nipy_resample.resample_nipy(nii_original,
-                                                target_resample,
+    nii_resampled = resample.resample_nipy(nii_original,
+                                           target_resample,
                                                  'mm', 'linear',
-                                                verbosity)
+                                           verbosity)
 
     nii_resampled = nipy2nifti(nii_resampled)
     pred_slices = segment_volume(nii_resampled, model_name, threshold,
@@ -331,10 +331,10 @@ def segment_file(input_filename, output_filename,
                                        volume_header)
     nii_segmentation = nifti2nipy(nii_segmentation)
 
-    nii_resampled_original = nipy_resample.resample_nipy(nii_segmentation,
-                                                         original_res,
+    nii_resampled_original = resample.resample_nipy(nii_segmentation,
+                                                    original_res,
                                                           'mm', 'linear',
-                                                         verbosity)
+                                                    verbosity)
     res_data = nii_resampled_original.get_data()
 
     # Threshold after resampling, only if specified
