+++ spinalcordtoolbox/resample.py
