+++ unit_testing/test_resample.py
@@ -11,7 +11,7 @@ import numpy as np
 import nibabel as nib
 from nipy.io.nifti_ref import nifti2nipy, nipy2nifti
 
-from spinalcordtoolbox.resample import nipy_resample
+from spinalcordtoolbox import resample
 
 
 @pytest.fixture(scope="session")
@@ -47,7 +47,7 @@ def fake_4dimage_nipy():
 # noinspection 801,PyShadowingNames
 def test_nipy_resample_image_3d(fake_3dimage_nipy):
     """Test resampling with 3D nipy image"""
-    img_r = nipy_resample.resample_nipy(fake_3dimage_nipy, new_size='2x2x1', new_size_type='factor', interpolation='nn')
+    img_r = resample.resample_nipy(fake_3dimage_nipy, new_size='2x2x1', new_size_type='factor', interpolation='nn')
     assert img_r.get_data().shape == (18, 18, 9)
     assert img_r.get_data()[8, 8, 4] == 1.0  # make sure there is no displacement in world coordinate system
     assert nipy2nifti(img_r).header.get_zooms() == (0.5, 0.5, 1.0)
@@ -58,7 +58,7 @@ def test_nipy_resample_image_3d(fake_3dimage_nipy):
 # noinspection 801,PyShadowingNames
 def test_nipy_resample_image_4d(fake_4dimage_nipy):
     """Test resampling with 4D nipy image"""
-    img_r = nipy_resample.resample_nipy(fake_4dimage_nipy, new_size='2x2x1x1', new_size_type='factor', interpolation='nn')
+    img_r = resample.resample_nipy(fake_4dimage_nipy, new_size='2x2x1x1', new_size_type='factor', interpolation='nn')
     assert img_r.get_data().shape == (18, 18, 9, 3)
     assert img_r.get_data()[8, 8, 4, 0] == 1.0  # make sure there is no displacement in world coordinate system
     assert img_r.get_data()[8, 8, 4, 1] == 0.0
